libcex: move the getifaddrs SIOCGIFAFLAG_IN6 FFI
Pure move of the BSD in6_ifreq FFI — the IN6_IFF_UNUSABLE flag mask, the In6Ifreq/In6Ifru request struct, and the siocgifaflag_in6 ioctl-code builder — into a BSD-gated libcex::in6_ifreq. Only visibility widened to pub(crate) (the union pad stays private); the size asserts and the request-code test travel with the items. interface/getifaddrs.rs references libcex::.
